Feature/review

+++ src/main/java/exambyte/application/dto/FrageDTO.java
@@ -7,6 +7,6 @@ import java.util.UUID;
 public record FrageDTO (
         UUID id,
         String frageText,
-        int maxPunkte,
+        double maxPunkte,
         UUID examId,
         QuestionTypeDTO type) {}
