+++ src/main/java/exambyte/application/service/ExamFacadeService.java
@@ -57,7 +57,7 @@ public interface ExamFacadeService {
 
     boolean antwortHasReview(AntwortDTO antwort);
 
-    void createReview(String bewertung, int punkte, UUID antwortId, UUID korrektorId);
+    void createReview(String bewertung, double punkte, UUID antwortId, UUID korrektorId);
 
     UUID getReviewerByName(String name);
