+++ src/main/java/exambyte/domain/export_mapper/ReviewExportDTOMapper.java
@@ -12,7 +12,7 @@ public interface ReviewExportDTOMapper {
 
     List<ReviewExportDTO> mapDTOToExport(ExamDTO exam,
                                    String reviewerName,
-                                   int maxPunkte,
+                                   double maxPunkte,
                                    List<FrageDTO> fragen,
                                    List<AntwortDTO> antworten,
                                    List<ReviewDTO> reviews);
